translated the constitution to Markdown
This change allows for easier modification. The TeX and PostScript versions
were deleted, as presumably you can build your own PDF copy if you like.

The unratified 2003 draft was also deleted in this change. Since the only
change was the addition of a Webmaster position (which we have no desire
to continue pursuing), there was no reason to keep it around.

Since this change did not alter the wording of the Constitution, no
ratification was required.